openid_connect_frontend.yaml.example: leverage <base_url> template
Setting an alternative issuer should not be an encouraged setup,
although provider discovery should work either way. The recommended
setting is to use the BASE as the issuer, and we can leverage the
agressive configuration value replacement logic, which rewrites all
occurences of <base_url> to the value of BASE. The unit test was
modified to guarantee this behaviour, though.